Added Dictionary.AddIndirect, PdfArrayy.AddIndirect
Also fix adding PdfFontType1Base14 properly
Also fix PdfVecObjects sorted handling
Also better SetDirty handling PdfObject/PdfDataContainerType
Also big cleaning

+++ src/podofo/base/PdfContentsTokenizer.cpp
@@ -63,9 +63,9 @@ PdfContentsTokenizer::PdfContentsTokenizer(PdfCanvas& pCanvas)
     if(contents->IsArray())
     {
         PdfArray& contentsArr = contents->GetArray();
-        for ( int i = 0; i < contentsArr.GetSize() ; i++ )
+        for (size_t i = 0; i < contentsArr.GetSize(); i++)
         {
-            auto &streamObj = contentsArr.FindAt(i);
+            auto& streamObj = contentsArr.FindAt(i);
             m_lstContents.push_back(&streamObj);
         }
     }
